+++ packages/playground-ui/src/services/mastra-runtime-provider.tsx
@@ -16,6 +16,7 @@ import { useWorkingMemory } from '@/domains/agents/context/agent-working-memory-
 import { MastraClient } from '@mastra/client-js';
 import { useAdapters } from '@/components/assistant-ui/hooks/use-adapters';
 import { MastraModelOutput } from '@mastra/core/stream';
+import { flushSync } from 'react-dom';
 
 const convertMessage = (message: ThreadMessageLike): ThreadMessageLike => {
   return message;
@@ -428,6 +429,45 @@ export function MastraRuntimeProvider({
                   break;
                 }
 
+                case 'tool-output': {
+                  if (!chunk.payload.output?.type.startsWith('workflow-')) return;
+
+                  flushSync(() => {
+                    setMessages(currentConversation => {
+                      const lastMessage = currentConversation[currentConversation.length - 1];
+                      const contentArray = Array.isArray(lastMessage.content)
+                        ? lastMessage.content
+                        : [{ type: 'text', text: lastMessage.content }];
+
+                      const newMessage = {
+                        ...lastMessage,
+                        content: contentArray.map(part => {
+                          if (part.type === 'tool-call') {
+                            return {
+                              ...part,
+                              ...chunk.payload,
+                              args: {
+                                ...part.args,
+                                __mastraMetadata: {
+                                  ...part.args?.__mastraMetadata,
+                                  partialChunk: chunk?.payload?.output,
+                                  isStreaming: true,
+                                },
+                              },
+                            };
+                          }
+
+                          return part;
+                        }),
+                      };
+
+                      return [...currentConversation.slice(0, -1), newMessage];
+                    });
+                  });
+
+                  break;
+                }
+
                 case 'tool-call': {
                   // Update the messages state
                   setMessages(currentConversation => {
@@ -446,7 +486,13 @@ export function MastraRuntimeProvider({
                                 type: 'tool-call',
                                 toolCallId: chunk.payload.toolCallId,
                                 toolName: chunk.payload.toolName,
-                                args: chunk.payload.args,
+                                args: {
+                                  ...chunk.payload.args,
+                                  __mastraMetadata: {
+                                    ...chunk.payload.args?.__mastraMetadata,
+                                    isStreaming: true,
+                                  },
+                                },
                               },
                             ]
                           : [
@@ -457,7 +503,13 @@ export function MastraRuntimeProvider({
                                 type: 'tool-call',
                                 toolCallId: chunk.payload.toolCallId,
                                 toolName: chunk.payload.toolName,
-                                args: chunk.payload.args,
+                                args: {
+                                  ...chunk.payload.args,
+                                  __mastraMetadata: {
+                                    ...chunk.payload.args?.__mastraMetadata,
+                                    isStreaming: true,
+                                  },
+                                },
                               },
                             ],
                       };
@@ -478,7 +530,10 @@ export function MastraRuntimeProvider({
                           type: 'tool-call',
                           toolCallId: chunk.payload.toolCallId,
                           toolName: chunk.payload.toolName,
-                          args: chunk.payload.args,
+                          args: {
+                            ...chunk.payload.args,
+                            __mastraMetadata: { ...chunk.payload.args?.__mastraMetadata, isStreaming: true },
+                          },
                         },
                       ],
                     };